Add unit tests and start SC update methodology
