+++ core/systemcontracts/upgrade.go
@@ -1,0 +1,23 @@
+package systemcontracts
+
+import (
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/state"
+)
+
+type UpgradeConfig struct {
+	BeforeUpgrade upgradeHook
+	AfterUpgrade  upgradeHook
+	ContractAddr  common.Address
+	CommitUrl     string
+	Code          string
+}
+
+type Upgrade struct {
+	UpgradeName string
+	Configs     []*UpgradeConfig
+}
+
+type upgradeHook func(blockNumber *big.Int, contractAddr common.Address, statedb *state.StateDB) error
